Updated CreateGame and Lobby
CreateGame and Lobby now no longer have to make get request to get their own user

Furthermore, timestamp is now set to creationTime of the server response
