+++ src/components/shared/routers/AppRouter.js
@@ -12,6 +12,7 @@ import SocialMode from "../../online/game/beforeGameStart/SocialMode";
 import Lobby from "../../online/game/beforeGameStart/Lobby";
 import CreateGame from "../../online/game/createGame/CreateGame"
 import Quickplay from "../../online/quickplay/Quickplay";
+import IntermediaryCounter from "../../online/game/beforeGameStart/IntermediaryCounter";
 /**
  * Main router of your application.
  * In the following class, different routes are rendered. In our case, there is a Login Route with matches the path "/offline"
@@ -82,6 +83,14 @@ class AppRouter extends React.Component {
                   render={() => (
                       <GameGuard>
                           <CreateGame/>
+                      </GameGuard>
+                  )}
+              />
+              <Route
+                  path="/intermediary"
+                  render={() => (
+                      <GameGuard>
+                          <IntermediaryCounter/>
                       </GameGuard>
                   )}
               />
